# docs/swagger.yaml
basePath: /
definitions:
  models.BoletoRequest:
    properties:
      beneficiario_agencia:
        type: string
      beneficiario_cnpj_cpf:
        example: 12.345.678/0001-90
        type: string
      beneficiario_conta:
        type: string
      beneficiario_endereco:
        type: string
      beneficiario_nome:
        example: Empresa Ltda
        type: string
      codigo_banco:
        example: 001-9
        type: string
      convenio:
        example: "12345"
        type: string
      data_vencimento:
        example: 31/12/2025
        type: string
      identificador_pagador:
        type: integer
      informacao_beneficiario:
        example: Multa de 2% após vencimento
        type: string
      nosso_numero:
        example: "123456"
        type: string
      numero_documento:
        example: "123456"
        type: string
      pagador_cpf:
        example: 123.456.789-00
        type: string
      pagador_endereco:
        type: string
      pagador_nome:
        example: João da Silva
        type: string
      parcela:
        example: 1
        type: integer
      retornar_pdf:
        example: true
        type: boolean
      valor:
        example: 150,75
        type: string
    required:
    - beneficiario_agencia
    - beneficiario_cnpj_cpf
    - beneficiario_conta
    - beneficiario_nome
    - codigo_banco
    - convenio
    - data_vencimento
    - informacao_beneficiario
    - nosso_numero
    - numero_documento
    - pagador_cpf
    - pagador_nome
    - parcela
    - valor
    type: object
  models.LinhaRetorno:
    properties:
      agencia_arrecadadora:
        type: string
      codigo:
        type: string
      codigo_barras:
        type: string
      curso_id:
        type: integer
      data_credito:
        type: string
      data_pagamento:
        type: string
      forma_arrecadacao:
        type: string
      forma_pagamento:
        type: string
      matricula_inscricao_por_extenso:
        type: string
      nsr:
        type: string
      numero_autenticacao:
        type: string
      parcela:
        type: integer
      tarifa:
        type: number
      valor_boleto:
        type: number
      valor_pago:
        type: number
    type: object
info:
  contact: {}
  description: API que recebe dados e gera um boleto em HTML.
  title: API de Geração de Boletos
  version: "1.0"
paths:
  /boleto:
    post:
      consumes:
      - application/json
      description: Retorna HTML ou PDF dependendo do parâmetro RetornarPDF
      parameters:
      - description: Dados do boleto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BoletoRequest'
      produces:
      - text/html
      - application/pdf
      responses:
        "200":
          description: PDF do boleto
          schema:
            type: file
        "400":
          description: Erro de validação
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Gera um boleto bancário
      tags:
      - boletos
  /retorno/read:
    post:
      consumes:
      - multipart/form-data
      description: Lê o conteúdo do arquivo de retorno bancário e extrai os boletos
      parameters:
      - description: Conteúdo do arquivo retorno (texto com múltiplas linhas)
        in: formData
        name: retorno
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LinhaRetorno'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Processa arquivo de retorno
      tags:
      - retorno
swagger: "2.0"